+++ frontend/src/commons/utilities/constants.ts
